Move custom GraphQL definitions to their own file
